+++ src/lib.rs
@@ -94,14 +94,12 @@
 mod error;
 mod istream;
 mod ostream;
-mod trim;
 mod types;
 mod varint;
 
 pub use error::{Error, Result};
 pub use istream::{IStream, Visitor};
 pub use ostream::{Flush, Handoff, Handover, NoFlush, NoHandoff, OStream, MIN_OUTPUT_BUFFER};
-pub use trim::{trim_tail, trim_tail_f32, trim_tail_f64};
 pub use types::{Id, Signed, Unsigned, API_VERSION, ID_MAX};
 
 #[cfg(feature = "fixlen")]